+++ app/src/main/java/com/ivione93/hojames/ui/trainings/AdapterGym.java
@@ -1,5 +1,6 @@
 package com.ivione93.hojames.ui.trainings;
 
+import android.app.AlertDialog;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -34,9 +35,17 @@ public class AdapterGym extends FirestoreRecyclerAdapter<Gym, AdapterGym.GymView
         }
 
         holder.ibDeleteGym.setOnClickListener(v -> {
-            db.collection("gym").document(model.id).delete();
-            notifyItemRangeChanged(position, getItemCount());
-            notifyItemRemoved(position);
+            AlertDialog.Builder deleteConfirm = new AlertDialog.Builder(v.getContext());
+            deleteConfirm.setTitle("Eliminar gimnasio");
+            deleteConfirm.setMessage("¿Está seguro que quiere eliminar el ejercicio?\n\nATENCIÓN: Se elimina sin necesidad de guardar el entrenamiento");
+            deleteConfirm.setCancelable(false);
+            deleteConfirm.setPositiveButton("Aceptar", (dialog, which) -> {
+                db.collection("gym").document(model.id).delete();
+                notifyItemRangeChanged(position, getItemCount());
+                notifyItemRemoved(position);
+            });
+            deleteConfirm.setNegativeButton("Cancelar", (dialog, which) -> dialog.cancel());
+            deleteConfirm.show();
         });
     }
