+++ app/src/main/java/com/ivione93/hojames/ui/trainings/AdapterSeries.java
@@ -1,5 +1,6 @@
 package com.ivione93.hojames.ui.trainings;
 
+import android.app.AlertDialog;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -30,9 +31,17 @@ public class AdapterSeries extends FirestoreRecyclerAdapter<Series, AdapterSerie
         holder.showTimeSerie.setText(Utils.getFormattedResult(model.time));
 
         holder.ibDeleteSerie.setOnClickListener(v -> {
-            db.collection("series").document(model.id).delete();
-            notifyItemRangeChanged(position, getItemCount());
-            notifyItemRemoved(position);
+            AlertDialog.Builder deleteConfirm = new AlertDialog.Builder(v.getContext());
+            deleteConfirm.setTitle("Eliminar serie");
+            deleteConfirm.setMessage("¿Está seguro que quiere eliminar la serie?\n\nATENCIÓN: Se elimina sin necesidad de guardar el entrenamiento");
+            deleteConfirm.setCancelable(false);
+            deleteConfirm.setPositiveButton("Aceptar", (dialog, which) -> {
+                db.collection("series").document(model.id).delete();
+                notifyItemRangeChanged(position, getItemCount());
+                notifyItemRemoved(position);
+            });
+            deleteConfirm.setNegativeButton("Cancelar", (dialog, which) -> dialog.cancel());
+            deleteConfirm.show();
         });
     }
